Parsed notes into "mbeat-readable" format
Right now, it's off by around 0.24 ms. Close enough to the original.
Parsed from the last beat instead of the first one. Things were a lot
easier after that. The whole thing's still pretty costly to process
(probably, I didn't make actual measurements) but it is not often
that the admin (me) would add songs so it's good enough.