[Fixed] URLs not working on HTTPS
When using HTTPS, assets and scripts fail to be loaded because they
use HTTP. URLs are generated using Laravel's asset() and url() hel-
per functions and they have a parameter to generate secure or unse-
cure URLs. The default is to generate unsecure URLs.

**Fix:** Passed the environment variable USE_HTTPS to asset() and
url() calls to set the security of generated URLs.